Rename twitchChannels to ChannelDropdown to follow naming of the container
